major updates change in format for the AudioNFTHexString.  Now all file info is in json format AudioNFTHexString in data section metadata.json files can be split with '@*-*@' AudioNFTHexString[0] = audioNFT ID starting AudioNFTHexString[1] = version number, currently 0.0.2 AudioNFTHexString[2] = all information needed to seperate Image/Audio File(s) AudioNfTHexString[2] can be converted directly from string to json. Additional updates and key pair information to be released soon
AudioNFT now supports attaching multiple audio files to a single image file.

AudioNFTIdString in Hex form is now the filename for a single metadata file.
the image collection number is appended to the end of the file name.
AudioNFT's can be identified on chain by the hex string in the file name.
the full AudioNFTIDHexString can be found in the data section of the metadata file

Future updates planned in order as follows:
code clean up (it's late I have to run to bed the code is a mess, but functional)
create script for image/audio extraction from AudioNFT files. all info needed to split files
is available in the AudioNFTHexString in the metadata.json data field.
create minting script.
Add additionalfunctionality as needed.   working towards playlist music player scripts.

currently changing file names from image format to audio format will cause the file to play in vlc and wmp